refactor: :recycle: Move formatUserProfilePath to ~/server/utils/users.
