+++ server/utils/files/index.ts
@@ -2,4 +2,3 @@ export * from "./formatFileName";
 export * from "./extractNamePart";
 export * from "./uploadedFiles";
 export * from "./createUploadedFileFromBase64";
-export * from "./formatUserProfilePath";
