+++ server/utils/users/index.ts
@@ -1,1 +1,2 @@
 export * from "./saveUserProfile";
+export * from "./formatUserProfilePath";
